114. Flatten Binary Tree to Linked List
Algorithm

We'll have a separate function for flattening out the tree since the main function provided in the problem isn't supposed to return anything and our algorithm will return the tail node of the flattened out tree.

For a given node, we will recursively flatten out the left and the right subtrees and store their corresponding tail nodes in leftTail and rightTail respectively.

Next, we will make the following connections (only if there is a left child for the current node, else the leftTail would be null)

 leftTail.right = node.right
 node.right = node.left
 node.left = None

Next we have to return the tail of the final, flattened out tree rooted at node. So, if the node has a right child, then we will return the rightTail, else, we'll return the leftTail.